+++ aasrp/helper/srp_data.py
@@ -47,7 +47,7 @@ def payout_amount_html(payout_amount: int) -> str:
 
     return (
         '<span class="srp-payout d-flex justify-content-end align-items-baseline">'
-        '<span class="srp-payout-tooltip"><span class="srp-payout-amount d-block cursor-pointer">'
+        '<span class="srp-payout-tooltip"><span class="srp-payout-amount d-block">'
         f"#payout_amount_localized#</span></span><sup>{payout_amount_ctc_icon}</sup></span>"
     )